Setup `nyc` to produce code coverage reports
This extends the public TypeScript configuration and sets it to check
just our source files. The `all` setting must be `false` otherwise the
tool crashes with an OOM because the test runner launches VS Code
instances. The coverage check is set to false because we'll enable
failures for reduced coverage later. The reports generated will go into
Azure DevOps (cobertura) and Codecov (lcov, which also produces HTML
output), and a summary is printed to the console.